Changed name of utils.js -> dag_utils.js and moved functions from macro.js -> dag_utils.js
